fix(mem-cache): resolve macos clang compilation errors and warnings

+++ src/mem/cache/tags/compressed_tags.cc
@@ -144,7 +144,8 @@ CompressedTags::findVictim(const CacheBlk::KeyType& key,
     const uint64_t offset = extractSectorOffset(key.address);
     for (const auto& entry : superblock_entries){
         SuperBlk* superblock = static_cast<SuperBlk*>(entry);
-        if (superblock->match(key) && !superblock->isLogicalMapped(offset) &&
+        if (superblock->match(key) &&
+            !superblock->isLogicalMapped(static_cast<int>(offset)) &&
             superblock->hasFreePhysicalSlot() && superblock->isCompressed() &&
             superblock->canCoAllocate(compressed_size)) {
             victim_superblock = superblock;
@@ -176,7 +177,8 @@ CompressedTags::findVictim(const CacheBlk::KeyType& key,
     }
 
     // Assign logical offset to a physical sub-block slot in victim superblock
-    int p_idx = victim_superblock->mapLogicalToPhysical(offset);
+    int p_idx =
+        victim_superblock->mapLogicalToPhysical(static_cast<int>(offset));
     SectorSubBlk *victim = victim_superblock->blks[p_idx];
 
     // It would be a hit if victim was valid in a co-allocation, and upgrades
